+++ RELEASE_NOTES.md
@@ -10,6 +10,8 @@
 - [GL] Cleanup up shader program creation and caching. Removed various methods for creating and compiling shaders from `Context`.
 - Removed obsolete IResourceManager interface
 - Removed various other obsolete APIs
+- [Vulkan] Removed unused `Pipeline` type
+- [Vulkan] Removed `TextureFormat.ofGLSLType`
 
 ### 5.4.10
 - [OpenVR] changed GL texture submit to 2 textures (previously side by side, issue with Quest 3)